packages/voice-backend/pyproject.toml
[project]
name = "iris-voice-backend"
version = "0.1.0"
description = "Voice processing backend for IRIS - STT (faster-whisper) + TTS (Chatterbox)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "faster-whisper>=1.1.0",
    "chatterbox-tts>=0.2.0",
    "numpy>=1.26.0",
    "scipy>=1.14.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]
gpu = [
    "torch>=2.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
